修改了calc_create_interf

+++ RealTimeAnalyse/PriorityArrangement.h
@@ -19,18 +19,18 @@ int offset_trans(int target, int basis, int T);
 bool find_interval(const std::vector<canfd_frame*>& frame_set, std::vector<int>& lower_bound, std::vector<int>& upper_bound);
 class betaset {
 public:
-    int C;   //执行时间
+    double C;   //执行时间
     int tr;  //release时间
-    betaset(int C,int tr):C(C),tr(tr){}
+    betaset(double C, int tr) :C(C), tr(tr) {}
 };
 //生成β集合
 bool create_beta(const std::vector<canfd_frame*>& frame_set, const canfd_frame& frame, int lower_bound, std::vector<betaset>& beta);
 //计算剩余干涉R_t_i：在时刻upper时 release的优先级 i 任务所面对的之前高优先级任务未完成的剩余执行时间,lower=upper-T_i+D_i
-int calc_remain_interf(const canfd_frame& frame, int t, std::vector<betaset>& beta);
+double calc_remain_interf(const canfd_frame& frame, int t, std::vector<betaset>& beta);
 //生成η集合
 bool  create_eta(const std::vector<canfd_frame*>& frame_set, const canfd_frame& frame, int t, int R, std::vector<betaset>& eta);
 //计算在t时刻优先级i 任务release后release的新任务产生的影响时间
-int calc_create_interf(const canfd_frame& frame, const int t, const int R, const std::vector<betaset>& eta);
+double calc_create_interf(const canfd_frame& frame, const int t, const int R, const std::vector<betaset>& eta);
 //检测任务集可行性
 //bool feasibility_check(std::vector<message>& messageSet);
 //frame_set为待分配优先级任务集合，向其中标号为taski的任务分配最低的优先级，pri没啥用
